requirements.txt
absl-py==2.0.0
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
asteroid-filterbanks==0.4.0
astor==0.8.1
astroid==2.15.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
audiostretchy==1.3.5
autopep8==1.6.0
av==10.0.0
Babel==2.13.1
bangla==0.0.2
bce-python-sdk==0.8.96
blinker==1.7.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
bokeh==3.3.1
boltons==23.1.1
Bottleneck==1.3.7
braceexpand==0.1.7
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.2.0
coqpit==0.0.17
ctranslate2==3.21.0
cycler==0.12.1
Cython==0.29.30
datasets==2.14.6
dateparser==1.1.8
decorator==4.4.2
demucs @ git+https://github.com/facebookresearch/demucs@e976d93ecc3865e5757426930257e200846a520a
dill==0.3.4
Distance==0.1.3
distro==1.8.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring-to-markdown==0.11
dora-search==0.1.12
editdistance==0.6.2
einops==0.6.1
encodec==0.1.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
faster-whisper==0.9.0
filelock==3.9.0
fire==0.5.0
flake8==6.0.0
Flask==2.3.3
Flask-Babel==2.0.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
g2p-en==2.1.0
g2pkk==0.1.2
g2pM==0.1.2.5
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
greenlet==3.0.1
grpcio==1.59.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
h5py==3.10.0
hangul-romanize==0.1.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.19.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
inflect==5.6.2
install==1.3.5
intervaltree==3.1.0
ipython==8.17.2
isort==5.12.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
jsonmerge==1.9.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
julius==0.2.7
k-diffusion==0.0.16
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.0
lameenc==1.6.3
lazy-object-proxy==1.9.0
lazy_loader==0.3
librosa==0.8.1
lightning==2.1.1
lightning-utilities==0.9.0
llvmlite==0.40.1
loguru==0.7.2
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mido==1.3.0
mock==5.1.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.12.2
nara-wpe==0.0.9
networkx==2.8.8
nltk==3.8.1
note-seq==0.0.3
num2words==0.5.13
numba==0.57.0
numpy==1.23.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.2
onnxruntime-gpu==1.16.2
openai==0.28.0
openai-whisper==20231106
OpenCC==1.1.7
opencc-python-reimplemented==0.1.7
opencv-python==4.8.1.78
openunmix==1.2.1
opt-einsum==3.3.0
optuna==3.4.0
packaging==23.1
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleaudio==1.1.0
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle-gpu==2.4.0
paddlesde==0.2.5
paddleslim==2.4.1
paddlespeech==1.4.1
paddlespeech-feat==0.1.0
pandas==1.5.3
parameterized==0.9.0
parso==0.8.3
pathos==0.2.8
pattern-singleton==1.2.0
Pillow==9.3.0
platformdirs==3.11.0
pluggy==1.0.0
pooch==1.8.0
portalocker==2.8.2
pox==0.3.3
ppdiffusers==0.19.3
ppft==1.7.6.7
praatio==5.1.1
pretty-midi==0.2.10
prettytable==3.9.0
primePy==1.3
proglog==0.1.10
prompt-toolkit==3.0.41
protobuf==3.20.0
psutil==5.9.6
pure-eval==0.2.2
pyannote.audio==3.0.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydocstyle==6.2.3
pydub==0.25.1
pyflakes==3.0.1
Pygments==2.16.1
pygtrie==2.5.0
pylint==2.17.0
pynndescent==0.5.10
pyparsing==3.1.1
pypinyin==0.44.0
pypinyin-dict==0.6.0
pyreadline3==3.4.1
pysbd==0.3.4
pytest-runner==6.0.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
pytoolconfig==1.2.5
pytorch-lightning==2.1.1
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
pywin32==306
pyworld==0.3.4
PyYAML==6.0.1
pyzmq==25.1.1
rarfile==4.1
referencing==0.30.2
regex==2023.10.3
requests==2.28.1
requests-mock==1.11.0
requests-oauthlib==1.3.1
resampy==0.4.2
resize-right==0.0.2
retrying==1.3.4
rich==13.6.0
rope==1.7.0
rpds-py==0.12.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sacrebleu==2.3.2
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.3
semver==3.0.2
sentencepiece==0.1.99
sentry-sdk==1.34.0
seqeval==1.2.2
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==0.5.15
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
submitit==1.5.1
swig==4.1.1.post0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
TextGrid==1.5
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.5.1
timer==0.2.2
ToJyutping==0.2.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torch==2.1.0+cu118
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.1.0+cu118
torchdiffeq==0.2.3
torchmetrics==1.2.0
torchsde==0.2.6
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.64.1
trainer==0.0.31
traitlets==5.13.0
trampoline==0.1.2
transformers==4.33.3
treetable==0.2.5
TTS==0.20.2
typeguard==2.13.3
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.7.0
umap-learn==0.5.4
Unidecode==1.3.7
urllib3==1.26.13
uvicorn==0.24.0.post1
visualdl==2.4.2
wandb==0.16.0
wcwidth==0.2.9
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.1
whatthepatch==1.0.4
whisperx @ git+https://github.com/m-bain/whisperx.git@4ece2369d73465a4e50417d0f0a45e262343703a
win32-setctime==1.1.0
xxhash==3.4.1
xyzservices==2023.10.1
yacs==0.1.8
yapf==0.32.0
yarl==1.9.2
zhon==2.0.2
